+++ app/upgrades.go
@@ -101,6 +101,7 @@ var upgradesList = []string{
 	"v5.5.0",
 	"v5.5.1",
 	"v5.5.2",
+	"v5.5.5",
 }
 
 // if there is an override list, use that instead, for integration tests