+++ contracts/test/ERC20toCW20PointerTest.js
@@ -1,8 +1,9 @@
 const {
-    setupSigners, deployErc20PointerForCw20, getAdmin, deployWasm, WASM, ABI, registerPointerForCw20, testAPIEnabled,
+    setupSigners, deployErc20PointerForCw20, getAdmin, deployWasm, WASM, ABI, registerPointerForERC20, testAPIEnabled,
     incrementPointerVersion
 } = require("./lib");
 const { expect } = require("chai");
+const BigNumber = require('bignumber.js');
 
 describe("ERC20 to CW20 Pointer", function () {
     let accounts;
@@ -45,7 +46,7 @@ describe("ERC20 to CW20 Pointer", function () {
             describe("validation", function () {
                 it("should not allow a pointer to the pointer", async function () {
                     try {
-                        await registerPointerForCw20(await pointer.getAddress());
+                        await registerPointerForERC20(await pointer.getAddress());
                         expect.fail(`Expected to be prevented from creating a pointer`);
                     } catch (e) {
                         expect(e.message).to.include("contract deployment failed");
@@ -137,6 +138,30 @@ describe("ERC20 to CW20 Pointer", function () {
                     const allowance = await pointer.allowance(owner, spender);
                     expect(Number(allowance)).to.equal(0);
                 });
+
+                it("approvals above uint128 max int should work", async function() {
+                    const owner = accounts[0].evmAddress;
+                    const spender = accounts[1].evmAddress;
+                    const maxUint128 = new BigNumber("0xffffffffffffffffffffffffffffffff", 16);
+                    const tx = await pointer.approve(spender, maxUint128.toFixed());
+                    await tx.wait();
+                    const allowance = await pointer.allowance(owner, spender);
+                    expect(allowance).to.equal(maxUint128.toFixed());
+
+                    // approving uint128 max int + 1 should work but only approve uint128
+                    const maxUint128Plus1 = maxUint128.plus(1);
+                    const tx128plus1 = await pointer.approve(spender, maxUint128Plus1.toFixed());
+                    await tx128plus1.wait();
+                    const allowance128plus1 = await pointer.allowance(owner, spender);
+                    expect(allowance128plus1).to.equal(maxUint128.toFixed());
+
+                    // approving uint256 should also work but only approve uint128
+                    const maxUint256 = new BigNumber("0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff", 16);
+                    const tx256 = await pointer.approve(spender, maxUint256.toFixed());
+                    await tx256.wait();
+                    const allowance256 = await pointer.allowance(owner, spender);
+                    expect(allowance256).to.equal(maxUint128.toFixed());
+                });
             });
 
             describe("transferFrom()", function () {
@@ -219,7 +244,11 @@ describe("ERC20 to CW20 Pointer", function () {
             account1: 3000000
         });
 
-        describe("Pointer Upgrade", function () {
+        // Pointer version is going to be coupled with seid version going forward (as in,
+        // given a seid version, it's impossible to have multiple versions of pointer).
+        // We need to recreate the equivalent of the following test once we have a framework
+        // for simulating chain-level upgrade.
+        describe.skip("Pointer Upgrade", function () {
             let newPointer;
 
             before(async function () {