+++ x/evm/handler.go
@@ -46,6 +46,12 @@ func NewProposalHandler(k keeper.Keeper) govtypes.Handler {
 			return HandleAddERCCW20PointerProposal(ctx, &k, c)
 		case *types.AddERCCW721PointerProposal:
 			return HandleAddERCCW721PointerProposal(ctx, &k, c)
+		case *types.AddCWERC20PointerProposal:
+			return HandleAddCWERC20PointerProposal(ctx, &k, c)
+		case *types.AddCWERC721PointerProposal:
+			return HandleAddCWERC721PointerProposal(ctx, &k, c)
+		case *types.AddERCNativePointerProposalV2:
+			return HandleAddERCNativePointerProposalV2(ctx, &k, c)
 		default:
 			return sdkerrors.Wrapf(sdkerrors.ErrUnknownRequest, "unrecognized evm proposal content type: %T", c)
 		}