+++ x/evm/types/codec.go
@@ -36,6 +36,9 @@ func RegisterInterfaces(registry codectypes.InterfaceRegistry) {
 		&AddERCNativePointerProposal{},
 		&AddERCCW20PointerProposal{},
 		&AddERCCW721PointerProposal{},
+		&AddCWERC20PointerProposal{},
+		&AddCWERC721PointerProposal{},
+		&AddERCNativePointerProposalV2{},
 	)
 	registry.RegisterImplementations(
 		(*sdk.Msg)(nil),